+++ src/car_app_manager/__init__.py
@@ -1,5 +1,5 @@
 """Car App Manager - manage apps on an Android car head unit over ADB."""
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
 APP_NAME = "Car App Manager"
 APP_NAME_AR = "مدير تطبيقات السيارة"